+++ cmd/veneur-prometheus/main.go
@@ -1,9 +1,12 @@
 package main
 
 import (
+	"crypto/tls"
+	"crypto/x509"
 	"flag"
 	"fmt"
 	"io"
+	"io/ioutil"
 	"math"
 	"net/http"
 	"regexp"
@@ -22,14 +25,23 @@ var (
 	interval          = flag.String("i", "10s", "The interval at which to query. Value must be parseable by time.ParseDuration (https://golang.org/pkg/time/#ParseDuration).")
 	ignoredLabelsStr  = flag.String("ignored-labels", "", "A comma-seperated list of label name regexes to not export")
 	ignoredMetricsStr = flag.String("ignored-metrics", "", "A comma-seperated list of metric name regexes to not export")
-	prefix            = flag.String("p", "", "A prefix to append to any metrics emitted. Do not include a trailing period.")
+	prefix            = flag.String("p", "", "A prefix to append to any metrics emitted. Include a trailing period. (e.g. \"myservice.\")")
 	statsHost         = flag.String("s", "127.0.0.1:8126", "The host and port — like '127.0.0.1:8126' — to send our metrics to.")
+
+	// mTLS params for collecting metrics
+	cert   = flag.String("cert", "", "The path to a client cert to present to the server. Only used if using mTLS.")
+	key    = flag.String("key", "", "The path to a private key to use for mTLS. Only used if using mTLS.")
+	caCert = flag.String("cacert", "", "The path to a CA cert used to validate the server certificate. Only used if using mTLS.")
 )
 
 func main() {
 	flag.Parse()
 
-	c, _ := statsd.New(*statsHost)
+	statsClient, _ := statsd.New(*statsHost)
+
+	if *prefix != "" {
+		statsClient.Namespace = *prefix
+	}
 
 	if *debug {
 		logrus.SetLevel(logrus.DebugLevel)
@@ -54,17 +66,15 @@ func main() {
 		}
 	}
 
+	httpClient := newHTTPClient(*cert, *key, *caCert)
+
 	ticker := time.NewTicker(i)
 	for _ = range ticker.C {
-		collect(c, ignoredLabels, ignoredMetrics)
-	}
-
-	if *prefix != "" {
-		c.Namespace = *prefix
+		collect(httpClient, statsClient, ignoredLabels, ignoredMetrics)
 	}
 }
 
-func collect(c *statsd.Client, ignoredLabels []*regexp.Regexp, ignoredMetrics []*regexp.Regexp) {
+func collect(httpClient *http.Client, statsClient *statsd.Client, ignoredLabels []*regexp.Regexp, ignoredMetrics []*regexp.Regexp) {
 	logrus.WithFields(logrus.Fields{
 		"stats_host":      *statsHost,
 		"metrics_host":    *metricsHost,
@@ -72,7 +82,7 @@ func collect(c *statsd.Client, ignoredLabels []*regexp.Regexp, ignoredMetrics []
 		"ignored_metrics": ignoredMetrics,
 	}).Debug("Beginning collection")
 
-	resp, err := http.Get(*metricsHost)
+	resp, err := httpClient.Get(*metricsHost)
 	if err != nil {
 		logrus.WithError(err).WithField("metrics_host", *metricsHost).Warn(fmt.Sprintf("Failed to collect metrics"))
 		return
@@ -86,7 +96,7 @@ func collect(c *statsd.Client, ignoredLabels []*regexp.Regexp, ignoredMetrics []
 			// We've hit the end, break out!
 			break
 		} else if err != nil {
-			c.Count("veneur.prometheus.decode_errors_total", 1, nil, 1.0)
+			statsClient.Count("veneur.prometheus.decode_errors_total", 1, nil, 1.0)
 			logrus.WithError(err).Warn("Failed to decode a metric")
 			break
 		}
@@ -100,13 +110,13 @@ func collect(c *statsd.Client, ignoredLabels []*regexp.Regexp, ignoredMetrics []
 		case dto.MetricType_COUNTER:
 			for _, counter := range mf.GetMetric() {
 				tags := getTags(counter.GetLabel(), ignoredLabels)
-				c.Count(mf.GetName(), int64(counter.GetCounter().GetValue()), tags, 1.0)
+				statsClient.Count(mf.GetName(), int64(counter.GetCounter().GetValue()), tags, 1.0)
 				metricCount++
 			}
 		case dto.MetricType_GAUGE:
 			for _, gauge := range mf.GetMetric() {
 				tags := getTags(gauge.GetLabel(), ignoredLabels)
-				c.Gauge(mf.GetName(), float64(gauge.GetGauge().GetValue()), tags, 1.0)
+				statsClient.Gauge(mf.GetName(), float64(gauge.GetGauge().GetValue()), tags, 1.0)
 				metricCount++
 			}
 		case dto.MetricType_SUMMARY:
@@ -114,14 +124,14 @@ func collect(c *statsd.Client, ignoredLabels []*regexp.Regexp, ignoredMetrics []
 				tags := getTags(summary.GetLabel(), ignoredLabels)
 				name := mf.GetName()
 				data := summary.GetSummary()
-				c.Gauge(fmt.Sprintf("%s.sum", name), data.GetSampleSum(), tags, 1.0)
-				c.Count(fmt.Sprintf("%s.count", name), int64(data.GetSampleCount()), tags, 1.0)
+				statsClient.Gauge(fmt.Sprintf("%s.sum", name), data.GetSampleSum(), tags, 1.0)
+				statsClient.Count(fmt.Sprintf("%s.count", name), int64(data.GetSampleCount()), tags, 1.0)
 				metricCount += 2 // One for sum, one for count, one for each percentile bucket
 
 				for _, quantile := range data.GetQuantile() {
 					v := quantile.GetValue()
 					if !math.IsNaN(v) {
-						c.Gauge(fmt.Sprintf("%s.%dpercentile", name, int(quantile.GetQuantile()*100)), v, tags, 1.0)
+						statsClient.Gauge(fmt.Sprintf("%s.%dpercentile", name, int(quantile.GetQuantile()*100)), v, tags, 1.0)
 						metricCount++
 					}
 				}
@@ -131,22 +141,22 @@ func collect(c *statsd.Client, ignoredLabels []*regexp.Regexp, ignoredMetrics []
 				tags := getTags(histo.GetLabel(), ignoredLabels)
 				name := mf.GetName()
 				data := histo.GetHistogram()
-				c.Gauge(fmt.Sprintf("%s.sum", name), data.GetSampleSum(), tags, 1.0)
-				c.Count(fmt.Sprintf("%s.count", name), int64(data.GetSampleCount()), tags, 1.0)
+				statsClient.Gauge(fmt.Sprintf("%s.sum", name), data.GetSampleSum(), tags, 1.0)
+				statsClient.Count(fmt.Sprintf("%s.count", name), int64(data.GetSampleCount()), tags, 1.0)
 				metricCount += 2 // One for sum, one for count, one for each histo bucket
 
 				for _, bucket := range data.GetBucket() {
 					b := bucket.GetUpperBound()
 					if !math.IsNaN(b) {
-						c.Count(fmt.Sprintf("%s.le%f", name, b), int64(bucket.GetCumulativeCount()), tags, 1.0)
+						statsClient.Count(fmt.Sprintf("%s.le%f", name, b), int64(bucket.GetCumulativeCount()), tags, 1.0)
 						metricCount++
 					}
 				}
 			}
 		default:
-			c.Count("veneur.prometheus.unknown_metric_type_total", 1, nil, 1.0)
+			statsClient.Count("veneur.prometheus.unknown_metric_type_total", 1, nil, 1.0)
 		}
-		c.Count("veneur.prometheus.metrics_flushed_total", metricCount, nil, 1.0)
+		statsClient.Count("veneur.prometheus.metrics_flushed_total", metricCount, nil, 1.0)
 	}
 }
 
@@ -183,4 +193,37 @@ func shouldExportMetric(mf dto.MetricFamily, ignoredMetrics []*regexp.Regexp) bo
 	}
 
 	return true
+}
+
+func newHTTPClient(certPath, keyPath, caCertPath string) *http.Client {
+	var caCertPool *x509.CertPool
+	var clientCerts []tls.Certificate
+
+	if certPath != "" {
+		clientCert, err := tls.LoadX509KeyPair(certPath, keyPath)
+		if err != nil {
+			logrus.WithError(err).Fatalf("error reading client cert and key")
+		}
+		clientCerts = append(clientCerts, clientCert)
+	}
+
+	if caCertPath != "" {
+		caCertPool = x509.NewCertPool()
+		caCert, err := ioutil.ReadFile(caCertPath)
+		if err != nil {
+			logrus.WithError(err).Fatalf("error reading ca cert")
+		}
+		caCertPool.AppendCertsFromPEM(caCert)
+	}
+
+	client := &http.Client{
+		Transport: &http.Transport{
+			TLSClientConfig: &tls.Config{
+				Certificates: clientCerts,
+				RootCAs:      caCertPool,
+			},
+		},
+	}
+
+	return client
 }