+++ sinks/splunk/splunk.go
@@ -374,9 +374,17 @@ func (sss *splunkSpanSink) makeHTTPRequest(req *http.Request, cancel func()) {
 		dec := json.NewDecoder(resp.Body)
 		err := dec.Decode(&parsed)
 		if err != nil {
-			sss.log.WithError(err).
-				WithField("http_status_code", resp.StatusCode).
-				Warn("Could not parse response from splunk HEC")
+			entry := sss.log.WithError(err).
+				WithFields(logrus.Fields{
+					"http_status_code": resp.StatusCode,
+					"endpoint":         req.URL.String(),
+				})
+			if sss.log.Level >= logrus.DebugLevel {
+				body, _ := ioutil.ReadAll(dec.Buffered())
+				entry = entry.WithField("response_body", string(body))
+			}
+			entry.Warn("Could not parse response from splunk HEC")
+
 			return
 		}
 		cause = "error"
@@ -393,7 +401,7 @@ func (sss *splunkSpanSink) makeHTTPRequest(req *http.Request, cancel func()) {
 			"hec_status_code":   parsed.Code,
 			"hec_response_text": parsed.Text,
 			"event_number":      parsed.InvalidEventNumber,
-		}).Error("Error response from Splunk HEC")
+		}).Warn("Error response from Splunk HEC. (Splunk restarts may cause transient errors).")
 	}
 	samples.Add(ssf.Count(failureMetric, 1, map[string]string{
 		"cause":       cause,
@@ -435,10 +443,14 @@ func (sss *splunkSpanSink) Ingest(ssfSpan *ssf.SSFSpan) error {
 		return err
 	}
 
-	// choose (1/spanSampleRate) spans for sampling if any spans
-	// have the traceID of 0 or are declared indicator spans, they
-	// will always be chosen, regardless of the sample rate.
-	if !ssfSpan.Indicator && ssfSpan.TraceId%sss.spanSampleRate != 0 {
+	// wouldDrop indicates whether this span would be dropped
+	// according to the sample rate. (1/spanSampleRate) spans
+	// will be kept. `wouldDrop` is marked on Splunk events
+	// below when it's true
+	wouldDrop := ssfSpan.TraceId%sss.spanSampleRate != 0
+
+	// indicator spans are never sampled
+	if wouldDrop && !ssfSpan.Indicator {
 		atomic.AddUint32(&sss.skippedSpans, 1)
 		return nil
 	}
@@ -451,9 +463,9 @@ func (sss *splunkSpanSink) Ingest(ssfSpan *ssf.SSFSpan) error {
 	}
 
 	serialized := SerializedSSF{
-		TraceId:        strconv.FormatInt(ssfSpan.TraceId, 10),
-		Id:             strconv.FormatInt(ssfSpan.Id, 10),
-		ParentId:       strconv.FormatInt(ssfSpan.ParentId, 10),
+		TraceId:        strconv.FormatInt(ssfSpan.TraceId, 16),
+		Id:             strconv.FormatInt(ssfSpan.Id, 16),
+		ParentId:       strconv.FormatInt(ssfSpan.ParentId, 16),
 		StartTimestamp: float64(ssfSpan.StartTimestamp) / float64(time.Second),
 		EndTimestamp:   float64(ssfSpan.EndTimestamp) / float64(time.Second),
 		Duration:       ssfSpan.EndTimestamp - ssfSpan.StartTimestamp,
@@ -462,6 +474,13 @@ func (sss *splunkSpanSink) Ingest(ssfSpan *ssf.SSFSpan) error {
 		Tags:           ssfSpan.Tags,
 		Indicator:      ssfSpan.Indicator,
 		Name:           ssfSpan.Name,
+	}
+
+	if wouldDrop {
+		// if we would have dropped this span, the trace is marked as "partial"
+		// this lets us readily search for indicator spans that have full traces
+		// we only mark indicator spans this way
+		serialized.Partial = &wouldDrop
 	}
 
 	event := &Event{
@@ -497,4 +516,5 @@ type SerializedSSF struct {
 	Tags           map[string]string `json:"tags"`
 	Indicator      bool              `json:"indicator"`
 	Name           string            `json:"name"`
+	Partial        *bool             `json:"partial,omitempty"`
 }